+++ cosmetics-web/app/views/secondary_authentication/method/new.html.erb
