+++ cosmetics-web/app/views/secondary_authentication/resend_code/new.html.erb
