+++ cosmetics-web/app/views/secondary_authentication/secondary_authentication/app.html.erb
