+++ cosmetics-web/app/views/secondary_authentication/secondary_authentication/sms.html.erb
